Refactor, move functions to new packages, sort them
