+++ pkg/testsuite/handlers_test.go
@@ -30,9 +30,9 @@ import (
 	"github.com/gogatekeeper/gatekeeper/pkg/authorization"
 	"github.com/gogatekeeper/gatekeeper/pkg/constant"
 	"github.com/gogatekeeper/gatekeeper/pkg/keycloak/config"
-	"github.com/gogatekeeper/gatekeeper/pkg/keycloak/proxy"
 	"github.com/gogatekeeper/gatekeeper/pkg/proxy/models"
 	"github.com/gogatekeeper/gatekeeper/pkg/proxy/session"
+	"github.com/gogatekeeper/gatekeeper/pkg/utils"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -58,7 +58,7 @@ func TestDebugHandler(t *testing.T) {
 
 func TestExpirationHandler(t *testing.T) {
 	cfg := newFakeKeycloakConfig()
-	uri := proxy.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.ExpiredURL)
+	uri := utils.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.ExpiredURL)
 	requests := []fakeRequest{
 		{
 			URI:          uri,
@@ -94,7 +94,7 @@ func TestOauthRequestNotProxying(t *testing.T) {
 func TestLoginHandlerDisabled(t *testing.T) {
 	cfg := newFakeKeycloakConfig()
 	cfg.EnableLoginHandler = false
-	loginURL := proxy.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LoginURL)
+	loginURL := utils.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LoginURL)
 	requests := []fakeRequest{
 		{URI: loginURL, Method: http.MethodPost, ExpectedCode: http.StatusNotImplemented},
 		{URI: loginURL, ExpectedCode: http.StatusMethodNotAllowed},
@@ -113,7 +113,7 @@ func TestLoginHandlerNotDisabled(t *testing.T) {
 
 func TestLoginHandler(t *testing.T) {
 	cfg := newFakeKeycloakConfig()
-	uri := proxy.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LoginURL)
+	uri := utils.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LoginURL)
 
 	testCases := []struct {
 		Name              string
@@ -226,7 +226,7 @@ func TestLoginHandler(t *testing.T) {
 func TestSkipOpenIDProviderTLSVerifyLoginHandler(t *testing.T) {
 	cfg := newFakeKeycloakConfig()
 	cfg.SkipOpenIDProviderTLSVerify = true
-	uri := proxy.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LoginURL)
+	uri := utils.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LoginURL)
 	requests := []fakeRequest{
 		{
 			URI:          uri,
@@ -290,7 +290,7 @@ func TestSkipOpenIDProviderTLSVerifyLoginHandler(t *testing.T) {
 //nolint:funlen
 func TestTokenEncryptionLoginHandler(t *testing.T) {
 	cfg := newFakeKeycloakConfig()
-	uri := proxy.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LoginURL)
+	uri := utils.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LoginURL)
 	// !! it must be here because of how test is written
 	cfg.EncryptionKey = testEncryptionKey
 
@@ -556,7 +556,7 @@ func TestTokenEncryptionLoginHandler(t *testing.T) {
 
 func TestLogoutHandlerBadRequest(t *testing.T) {
 	cfg := newFakeKeycloakConfig()
-	logoutURL := proxy.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LogoutURL)
+	logoutURL := utils.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LogoutURL)
 
 	testCases := []struct {
 		Name              string
@@ -607,7 +607,7 @@ func TestLogoutHandlerBadRequest(t *testing.T) {
 func TestLogoutHandlerBadToken(t *testing.T) {
 	cfg := newFakeKeycloakConfig()
 	cfg.NoRedirects = true
-	logoutURL := proxy.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LogoutURL)
+	logoutURL := utils.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LogoutURL)
 	requests := []fakeRequest{
 		{
 			URI:          logoutURL,
@@ -633,7 +633,7 @@ func TestLogoutHandlerBadToken(t *testing.T) {
 
 func TestLogoutHandlerGood(t *testing.T) {
 	cfg := newFakeKeycloakConfig()
-	logoutURL := proxy.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LogoutURL)
+	logoutURL := utils.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LogoutURL)
 	testCases := []struct {
 		Name              string
 		ProxySettings     func(c *config.Config)
@@ -655,7 +655,7 @@ func TestLogoutHandlerGood(t *testing.T) {
 			ProxySettings: func(c *config.Config) {},
 			ExecutionSettings: []fakeRequest{
 				{
-					URI:              proxy.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LogoutURL) + "?redirect=http://example.com",
+					URI:              utils.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LogoutURL) + "?redirect=http://example.com",
 					HasToken:         true,
 					ExpectedCode:     http.StatusSeeOther,
 					ExpectedLocation: "http://example.com",
@@ -669,7 +669,7 @@ func TestLogoutHandlerGood(t *testing.T) {
 			},
 			ExecutionSettings: []fakeRequest{
 				{
-					URI:              proxy.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LogoutURL),
+					URI:              utils.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LogoutURL),
 					HasToken:         true,
 					ExpectedCode:     http.StatusSeeOther,
 					ExpectedLocation: "http://127.0.0.1",
@@ -683,7 +683,7 @@ func TestLogoutHandlerGood(t *testing.T) {
 			},
 			ExecutionSettings: []fakeRequest{
 				{
-					URI:          proxy.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LogoutURL) + "?redirect=",
+					URI:          utils.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LogoutURL) + "?redirect=",
 					HasToken:     true,
 					ExpectedCode: http.StatusSeeOther,
 				},
@@ -710,7 +710,7 @@ func TestLogoutHandlerGood(t *testing.T) {
 func TestSkipOpenIDProviderTLSVerifyLogoutHandler(t *testing.T) {
 	cfg := newFakeKeycloakConfig()
 	cfg.SkipOpenIDProviderTLSVerify = true
-	logoutURL := proxy.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LogoutURL)
+	logoutURL := utils.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LogoutURL)
 	requests := []fakeRequest{
 		{
 			URI:          logoutURL,
@@ -749,7 +749,7 @@ func TestSkipOpenIDProviderTLSVerifyLogoutHandler(t *testing.T) {
 func TestRevocation(t *testing.T) {
 	cfg := newFakeKeycloakConfig()
 	cfg.RevocationEndpoint = ""
-	logoutURL := proxy.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LogoutURL)
+	logoutURL := utils.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.LogoutURL)
 	requests := []fakeRequest{
 		{
 			URI:          logoutURL,
@@ -784,7 +784,7 @@ func TestRevocation(t *testing.T) {
 func TestTokenHandler(t *testing.T) {
 	cfg := newFakeKeycloakConfig()
 	cfg.NoRedirects = true
-	uri := proxy.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.TokenURL)
+	uri := utils.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.TokenURL)
 	goodToken, err := NewTestToken("example").GetToken()
 	if err != nil {
 		t.Fatalf("Error when creating test token %v", err)
@@ -884,7 +884,7 @@ func TestServiceRedirect(t *testing.T) {
 func TestAuthorizationURLWithSkipToken(t *testing.T) {
 	cfg := newFakeKeycloakConfig()
 	cfg.SkipTokenVerification = true
-	uri := proxy.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.AuthorizationURL)
+	uri := utils.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.AuthorizationURL)
 	newFakeProxy(cfg, &fakeAuthConfig{}).RunTests(t, []fakeRequest{
 		{
 			URI:          uri,
@@ -1090,7 +1090,7 @@ func TestAuthorizationURL(t *testing.T) {
 
 func TestCallbackURL(t *testing.T) {
 	cfg := newFakeKeycloakConfig()
-	uri := proxy.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.CallbackURL)
+	uri := utils.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.CallbackURL)
 	requests := []fakeRequest{
 		{
 			URI:          uri,
@@ -1125,7 +1125,7 @@ func TestCallbackURL(t *testing.T) {
 
 func TestHealthHandler(t *testing.T) {
 	cfg := newFakeKeycloakConfig()
-	uri := proxy.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.HealthURL)
+	uri := utils.WithOAuthURI(cfg.BaseURI, cfg.OAuthURI)(constant.HealthURL)
 	requests := []fakeRequest{
 		{
 			URI:          uri,
